refactor and add recurring event update functionalities
