+++ src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts
@@ -2,6 +2,7 @@
    Recurrence utility functions
 */
 
+import dayjs from 'dayjs';
 import {
   Days,
   dayNames,
@@ -10,7 +11,11 @@ import {
   weekDayOccurences,
 } from './recurrenceConstants';
 import { Frequency } from './recurrenceTypes';
-import type { WeekDays, InterfaceRecurrenceRuleState } from './recurrenceTypes';
+import type {
+  WeekDays,
+  InterfaceRecurrenceRuleState,
+  InterfaceRecurrenceRule,
+} from './recurrenceTypes';
 
 // function that generates the recurrence rule text to display
 // e.g. - 'Weekly on Sunday, until Feburary 23, 2029'
@@ -143,4 +148,105 @@ export const isLastOccurenceOfWeekDay = (date: Date): boolean => {
   }
 
   return date.getDate() === lastOccurenceInMonth.getDate();
+};
+
+// function that evaluates whether the startDate or endDate of a recurring event instance have changed
+export const haveInstanceDatesChanged = (
+  instanceOriginalStartDate: string,
+  instanceOriginalEndDate: string,
+  instanceNewStartDate: string,
+  instanceNewEndDate: string,
+): boolean => {
+  return (
+    instanceOriginalStartDate !== instanceNewStartDate ||
+    instanceOriginalEndDate !== instanceNewEndDate
+  );
+};
+
+// function that checks whether the recurrence rule has changed
+export const hasRecurrenceRuleChanged = (
+  originalRecurrencerule: InterfaceRecurrenceRule | null,
+  recurrenceRuleState: InterfaceRecurrenceRuleState,
+): boolean => {
+  if (!originalRecurrencerule) {
+    return false;
+  }
+
+  const newRecurrenceRule = getRecurrenceRule(recurrenceRuleState);
+
+  const recurrenceProperties = Object.keys(
+    newRecurrenceRule,
+  ) as (keyof InterfaceRecurrenceRule)[];
+
+  for (const recurrenceProperty of recurrenceProperties) {
+    if (recurrenceProperty === 'weekDays') {
+      if (
+        weekDaysHaveChanged(
+          originalRecurrencerule.weekDays,
+          newRecurrenceRule.weekDays,
+        )
+      ) {
+        return true;
+      }
+    } else if (
+      originalRecurrencerule[recurrenceProperty] !==
+      newRecurrenceRule[recurrenceProperty]
+    ) {
+      return true;
+    }
+  }
+
+  return false;
+};
+
+// function that returns the recurrence rule object based on the current recurrence rule state
+const getRecurrenceRule = (
+  recurrenceRuleState: InterfaceRecurrenceRuleState,
+): InterfaceRecurrenceRule => {
+  const {
+    recurrenceStartDate,
+    recurrenceEndDate,
+    frequency,
+    weekDays,
+    interval,
+    count,
+    weekDayOccurenceInMonth,
+  } = recurrenceRuleState;
+
+  const originalRecurrencerule = {
+    recurrenceStartDate: dayjs(recurrenceStartDate).format('YYYY-MM-DD'),
+    recurrenceEndDate: recurrenceEndDate
+      ? dayjs(recurrenceEndDate).format('YYYY-MM-DD')
+      : null,
+    frequency,
+    weekDays: weekDays?.length ? weekDays : [],
+    interval,
+    count: count ?? null,
+    weekDayOccurenceInMonth: weekDayOccurenceInMonth ?? null,
+  };
+
+  return originalRecurrencerule;
+};
+
+// function to check whether recurrence weekDays have been changed
+const weekDaysHaveChanged = (
+  originalWeekDays: WeekDays[],
+  currentWeekDays: WeekDays[],
+): boolean => {
+  if (originalWeekDays.length !== currentWeekDays.length) {
+    return true;
+  }
+
+  // Sort both arrays
+  const sortedOriginalWeekDays = [...originalWeekDays].sort();
+  const sortedCurrentWeekDays = [...currentWeekDays].sort();
+
+  // Compare arrays
+  for (let i = 0; i < sortedOriginalWeekDays.length; i++) {
+    if (sortedOriginalWeekDays[i] !== sortedCurrentWeekDays[i]) {
+      return true;
+    }
+  }
+
+  return false;
 };
